add Future/Or call() extensions, remove all ~ ops
I think this should be the final breaking release. Everything is consistent across the library now and we finally have a way to cleanly unwrap Option/Result values that allows us to access properties on the resulting value without having to do any precedence juggling (except in the cases of awaited values but that's the case with any future, not just option/result futures).

I just wish I had thought of this sooner instead of going the operator route and not realizing how problematic it was until actually incorporating the lib into the project I'm working on.